src/requirements.txt
requests==2.31.0
urllib3==2.0.7
fastapi==0.109.1
httpx==0.24.1
uvicorn[standard]==0.22.0
gunicorn==20.1.0
pytest-cov==4.1.0
pytest-dotenv==0.5.2
python-multipart==0.0.6
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.17.0
fastapi-health==0.4.0
python-decouple==3.8
ecs-logging==2.0.2
opencv-contrib-python==4.7.0.72
pytesseract==0.3.10
Pillow>=10.0.0
pydantic<=2.0
parameterized==0.9.0
networkx==2.5.1
shapely==2.0.1
tqdm==4.65.0
matplotlib==3.7.2
pyodbc==4.0.39
pdf2image
pypdfium2